Allfiles/Labs/02/Starter/API/requirements.txt
flask
gunicorn
azure-storage-blob
requests

